renames graphql schema file
