+++ superai/apis/meta_ai/model.py
@@ -6,7 +6,7 @@ from .session import MetaAISession
 
 log = logger.get_logger(__name__)
 
-from superai.apis.meta_ai.meta_ai_schema import (
+from superai.apis.meta_ai.meta_ai_graphql_schema import (
     meta_ai_app_constraint,
     meta_ai_app_insert_input,
     meta_ai_app_on_conflict,
